Bug fixes related to the Journal#shutdownNow() deadlock.  Several problems were identified, including:
NonBlockingLockManagerWithNewDesign: When a task came due for execution, a RejectedExecutionException was not being trapped and set on the task.  This caused the caller (which in the case of the shutdown deadlock was waiting on a SinglePhaseCommit task) to wait forever rather than noticing the RejectedExecutionException.

WriteExecutorService: This was modified to override shutdown() and shutdownNow() to shutdown the embedded lock manager instance.

Changes to Journal and JournalTransactionService are cosmetic only (final declaration for a method argument, etc).

The TestJournalTxShutdownAndRestart class was removed since the tests were never written and the suite was not being executed.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@2601 8f7bc0f5-282e-0410-95e3-8d296e9bb460