I have hooked up AbstractJournal for "HA" with itself.  It follows the basic pattern of delegating requests for commitNow() and abort() to the quorum, which them applies them to its HAGlue object.  For true HA, the quorum implementation needs to apply the methods to the HAGlue object for each of the failover nodes as well, which this does not do.  It also does not have a hook for _bufferStrategy.truncate(long extent) and it does not have hooks to setup the write service cache for the quorum.  I have run the unit tests against the WORMStrategy for the journal and it all looks good.
I've added a pattern where the quorum token (based on the commitTime at which the quorum "met") is used to verify that the operation may proceed.  This provides a nice check against changes in the constituted quorum without worrying about the specifics of the quorum state in any detail.

Maybe tomorrow we can hook this up to the buffer strategy (for truncate) and write service cache (for the write pipeline)?  Also, I think that handling of bad reads might well go through the resource service per your modifications, so I have left it off the HAGlue API for now.  I also have not explored how to export the HAGlue interface as a proxy, but that is what it should be.

For scale-out HA we will have to add various methods such as closeForWrites() and deal with the fact that the journal changes on each overflow, but this feels like a decent start.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@2674 8f7bc0f5-282e-0410-95e3-8d296e9bb460