Cleaned up WORMStrategy by removing the older WriteCache integration.
Renamed WriteCacheService2 to WriteCacheService (removing the older version of WriteCacheService).

Added options to FileMetadata for the #of write cache buffers and whether or not record-level checksums are enabled (they are as of root block version ONE for the RW store and root block version TWO for the WORM store).

Added an Option for the #of write cache buffers.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@2701 8f7bc0f5-282e-0410-95e3-8d296e9bb460