Some progress.
The "optimization" for the checksums is pretty leaky, which is the source of most of the problems.  It shows up in the maximum record size for the store, in different code paths for reading or writing on the store, etc., etc.  Also, I found a problem with assertEquals() for
ByteBuffers.  When limit() != capacity() it was assuming that it could compare the backing arrays, but that is not true when we tack on the 4 bytes for the checksum.  I have consolidated the versions of that assertEquals variant into TestCase3 (extends TestCase2).

TestWORMStrategy is now mostly Ok.

test_writeNoExtend and test_writeWithExtend fail due to the additional
4 bytes required when we use checksums.  

TestAbort#test_abort() fails due to some issues with writing and
reading records larger than the cache.  It fails whether or not
checksums are enabled.  This is one of those code path things. 

StressTestConcurrentTx WORMStrategy w/o checksums can deadlock on
a moderately reliable basis.  I have not seen this yet when running
with w/ checksums.

I am currently enabling or disabling checksums for WORMStrategy in
the constructor, but that is just while exploring these problems.  It
is automatically enabled in the committed code for new WORM stores.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@2704 8f7bc0f5-282e-0410-95e3-8d296e9bb460