Modified the HASendService to only support "incremental" send and renamed the "incSend()" method as send.  You must always get() the Future returned by send() to await the outcome of the operation.
Modified the HAReceiveService to only open the client connection (both reading from upstream and writing on downstream) once.  Those connections are held open in order to avoid both latency with re-establishing a client connection and (more critically) errors which can be thrown if there are attempts to close/open a socket rapidly.  While this will often succeed, it can fail.  The present arrangement is more robust.

The upstream "client" connection in the HAReceiveService was encapsulated by a "Client" class in order to simplify some life cycle management issues.

The unit tests were updated to use "incSend()" semantics.

All is good.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@2842 8f7bc0f5-282e-0410-95e3-8d296e9bb460