Merge from branch to trunk [r2895:r2912] (star joins).
This has only been tested against triples at this point.  The star join branch is being merged into the trunk so we can modify the join task to allow concurrent execution of the access path tasks, which will translate into concurrent star joins.  It is anticipated that this will accelerate star joins tremendously.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/starjoin@2913 8f7bc0f5-282e-0410-95e3-8d296e9bb460