Moved the services (queue) counters for the rdf data load analysis into a separate path to reduce the clutter.
Made the informational message about delete index segment files INFO rather than WARN since it was cluttering the detail logs.

Modified the event detail for an index segment build to clearly indicate MB(out) versus MB(in).

Modified the Latch to propagate an InterruptedException (or a nested InterruptedException) onto the caller's thread.  I doubt that this is responsible for the client failure to progress since it would have logged an error anyway and no errors were reported in the logs.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@3052 8f7bc0f5-282e-0410-95e3-8d296e9bb460