I have implemented a unicode clean option for encoding the schema name in the SparseRowStore keys.  This option is currently disabled by default, which provides backward compatibility.
Note that I have not been able to generate a schema name which in fact caused the JDK collation rules to embed a nul byte into the key.  I imagine that the constraints on the legal patterns for schema names preclude many cases which might otherwise have caused a problem.  However, I would not be surprised to learn that legal schema names could be used to generate Unicode sort keys with embedded nul bytes using the JDK CollatorEnum option.

I have left the unicode clean option disabled for the moment so we can reflect on the best way to handle this.  For example, if we put the bigdata release version number into the code and from the code into the persistence store, then we could automatically detect the version of the code used to create a given persistent data structure.  Something along these lines could facilitate automatic data migration.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@3367 8f7bc0f5-282e-0410-95e3-8d296e9bb460