Merge trunk to branch [r3391:r3437].
Note: The edits by BrianM to fix the test data URIs are at least partially missing in the HA branch.  Therefore we need to reconcile the branch against the trunk in depth in the file system (winmerge, ediff) before merging from the HA branch back into the trunk. It looks like these changes should have been introduced in r2599, which is before just the start of the HA branch, and possibly r3305.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/JOURNAL_HA_BRANCH@3438 8f7bc0f5-282e-0410-95e3-8d296e9bb460