I raised the getFromKey() and getToKey() methods onto IKeyOrder, defined an AbstractKeyOrder, made the various existing implementations derive from that, and modified SPOKeyOrder to override a protected inner method which handles the encoding of a single component of the key so it can do the right thing with the IVs.  I moved the SPOAccessPath#init() logic into AbstractAccessPath, which makes the fromKey and toKey final and AbstractAccessPath is no longer abstract and could be renamed at some point.
I have setup some unit tests in com.bigdata.bop.ap where I am starting to work through testing bop evaluation in the context of a simple relation "R" and relation element type "E" which are defined in that package in the test suite.  See TestPredicateAccessPath in that package in the test suite.

I would like to get this to the point where we are using the annotated bops as we plan to use them for scale up/out query.

I would like to test the ability to use bops to describe filters (IElementFilter) and constraints (IConstraint) on IPredicates. 

I plan to add IConstraint's to IPredicate.  We will renaming IPredicate#getConstraint():IElementFilter as getElementFilter() and add an IPredicate#getConstraint():IConstraint method.  We can use the IConstraint on the IPredicate to validate what used to verify "rule" level constraints and this will let us customize which constraints we verify where (if we need to do that).

I also want to use an optional "variablesToKeep" annotation do things like stripping off undesired variables (when it is not specified we pass along all variables).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3452 8f7bc0f5-282e-0410-95e3-8d296e9bb460