Added a class (DirectBufferPoolAllocator) to manage allocations of direct ByteBuffer slices against the DirectBufferPool for use with the FederatedQueryEngine. There is a limited test suite for this class.
Partly integrated the DirectBufferPoolAllocator with the FederatedQueryEngine.  The generated intermediate results are now written onto direct ByteBuffer slices allocated on the DirectBufferPool and notification messages are sent to the receiving services.  I have not yet modified the receiving FederatedQueryEngine to demand the data from the remote query engine.

Harmonized InsertOp and PipelineJoin as IShardwiseBindingSetOps. This makes it possible to find the predicate for the access path on which we need to read/write and map the results to the appropriate shards.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3524 8f7bc0f5-282e-0410-95e3-8d296e9bb460