Further reorganization of the federated query engine and its use of buffers.  There are now thick (payload including with RMI) and thin (RMI message with payload via NIO over the resource service) messages for moving chunks around during distributed query processing. There are bare bones unit tests for these as well.  (The ResourceServer does not actually use NIO yet, but it can be optimized later.  Also, we are not yet applying compression suitable for binding sets, but again that can be an optimization.)
Now that RMI messages and payload transfers are more or less in place, I am going to work through some unit tests of distributed query evaluation.  To do that I still need to reconcile the concept of a "query buffer" where the final solutions are written with the new model for moving data around.  I think that the query buffer will no longer be privledged (it used to be a proxy object for a buffer on the client).  Instead, scale-out will require an operator at the top of the query plan whose evaluation context is the query controller.  The mere presence of an operator which copies its inputs to its outputs whose evaluation context is the query controller is sufficient to do the trick.  For practical purposes, this can be a SliceOp, since that already must run in the query controller context.  If an offset/limit are not specified, then they can be set to 0L and MAX_LONG on the SliceOp which has the effect of turning it into a NOP (unless you are visiting an unbelievable #of results).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3532 8f7bc0f5-282e-0410-95e3-8d296e9bb460