Refactored the RunState and ChunkTask out of the RunningQuery.
Working on the federation based unit test setup.  We can not use the EmbeddedFederation for this because the serviceId is shared by both data service instances.  Unfortunately, we can no longer easily use the JiniServiceHelper either due to things like the jini group setup.  I am going to tackle this next on a workstation with more RAM so I can attach to a running federation.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3543 8f7bc0f5-282e-0410-95e3-8d296e9bb460