Branch created for https://sourceforge.net/apps/trac/bigdata/ticket/166.
Implement a feature for reporting the change set via the BigdataSail. The implementation will report on the delta in the statements materialized in database between the previous commit point and the current commit point. This report will capture, as of the new commit point: (1) all given statements written onto the database, if they were not previously present in the database, and all inferences materialized as a result of the statements asserted, if they were not already present in the database; and (2) all given statements which were retracted from the database, if they can no longer be proven based on the remaining statements, and all materialized inferences which are no longer proven, given the statements which were retracted, and which were therefore retracted from the database.

In order to signal the intention of the application, it might be best to create a second "commit()" method, e.g., named "commitAndReportChangeSet()", which provides the additional reporting mechanisms.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/CHANGE_SET_BRANCH@3609 8f7bc0f5-282e-0410-95e3-8d296e9bb460