Added support for RMI access paths.  All operators now assume "ANY" as their evaluation context.  This means that you must explicitly override the evaluation context for scale-out JOINS in order to use shard-partitioned access paths versus remote access paths.
The default of "ANY" is not valid for some operators.  Such operators now check in their shallow copy constructor to verify that a legal evaluation context was explicitly set.

At this point we have the tools to evaluate a default graph query in scale-out.  The Rule2BOpUtility needs to be modified to become aware of standalone versus scale-out and to use the appropriate operator patterns for standalone and scale-out.  The correct operator pattern depends on a number of factors.  I will be working up a cost model and pulling the logic out of DefaultGraphSolutionExpander and NamedGraphSolutionExpander.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3630 8f7bc0f5-282e-0410-95e3-8d296e9bb460