Reorganized the pipeline operators, the index sampling operators, and the CTC style striterator operators.
Deprecated IPredicate#CONSTRAINTS.  It will be replaced by INDEX_LOCAL_FILTERS and ACCESS_PATH_FILTERS.  The former will always run locally at the index against the ITupleIterator.  The latter will always run on the AccessPath side against the materialized elements.  For example, the default graph "distinct spo" constraint will be imposed as a DistinctFilter wrapping a StripContextFilter.  In scale-out, this will use a remote access path as necessary in order to have a total view of the sharded data.  My next step is to implement that distinction.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3665 8f7bc0f5-282e-0410-95e3-8d296e9bb460