Added some interfaces for reporting on those aspects of a B+Tree which feed into cost estimates for query plans.
Integrated the named graph decision tree logic with the B+Tree cost model in QueryEngine. I've only done this for standalone as this point.  We will have to override this for scale-out to use an approximation of the mixtures of cost models for a BTree on the Journal and ~3 index segment files.  That can be overridden in FederatedQueryEngine.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3713 8f7bc0f5-282e-0410-95e3-8d296e9bb460