Resolution for https://sourceforge.net/apps/trac/bigdata/ticket/185.
The zookeeper locks are not reentrant.  Since the full text index is in the namespace of the lexicon relation, the full text index must not attempt to acquire the zlock for that relation when destroying the full text index.  FullTextIndex#destroy() has been overridden in BigdataRDFFullTextIndex to avoid attempting to acquire a non-reentrant lock already held by the caller.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3805 8f7bc0f5-282e-0410-95e3-8d296e9bb460