Added a MultiplexBlockingBuffer.  This is a factory pattern which may be used to share the same backing BlockingBuffer among many producers.  Each producer receives a skin for the backing buffer. The backing buffer is only closed once each producer closes their skin.
Added an IMultiSourceAsynchronousIterator interface for an IAsynchronousIterator which can consume multiple sources.  There is one implementation in this commit, which allows the producer to attach another source.  This is used to assign a chunk to a task which is already running.  There is another version which handles multiple concurrent producers, but it's implementation is not yet finished.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3830 8f7bc0f5-282e-0410-95e3-8d296e9bb460