Modified BOpContext to support IMultiSourceAsynchronousIterator so we can attach sources to already running tasks.
Modified BOpContextBase to hold a hard reference to the Executor to avoid problems with errors reported up from the IIndexManager if it has been closed.  Since the caller now has access to the Executor after the IIndexManager is closed, the relevant code in RunningQuery now sees a RejectedExecutionException rather than an IllegalStateException.

Modified RunningQuery to attach new chunks to already running tasks, at least in standalone.  There is more work that needs to be done here which falls broadly under the category of performance optimizations of the query engine.  This optimization is not yet available in scale-out because an RMI is necessary back to the controller and that should not happen in the QueryEngine's run Thread.  

There is a known problem with high volume queries, such as LUBM Q9 on U10 or above, that they can block.  What appears to be happening is that a join is running into a bounded queue (a BlockingBuffer with a limited capacity).  The code needs to be modified either to use an unbounded queue (potentially backed by a direct ByteBuffer), or to emit multiple IChunkMessages (this option was historically used in scale-out), or to chain the consumers and producers together as we did historically in the trunk (this option is very efficient in standalone).

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3834 8f7bc0f5-282e-0410-95e3-8d296e9bb460