Reviewing the changes to the NonBlockingLockManagerWithNewDesign, it seems to me that line 2100 should be:
// One more time through the loop to exit @ Halted. 
continue;
 
Rather than:

// Done.
awaitStateChange(ShutdownNow);

The awaitStateChange(ShutdownNow) call at 2100 in your edit will return immediately since the expected run state was just set to Halted a few lines above in the file.   It is as far as I can tell a NOP.  However, the "continue" will reenter the loop at the top, observe the Halted run state, and then execute the case for "Halted" which shuts down the service running the AcceptTask.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3859 8f7bc0f5-282e-0410-95e3-8d296e9bb460