RunningQuery - Modified to not log an InterruptedException for an operator task @ ERROR since this is the normal behavior of a SliceOp when its LIMIT is satisfied.
TestQueryEngine - Modified a unit test which uses SliceOp to limit the #of results visited to not check the statistics on the join operator used in the query.  The statistics of the join operator are not reliably updated because there is a race condition between the SliceOp, which interrupts the join task, and the join task's normal completion and post-processing.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3861 8f7bc0f5-282e-0410-95e3-8d296e9bb460