Added a sumRangeCounts counter to the join stats.
Modified the runtime optimizer to use the sum of the range counts considered by the cutoff join when the cardinality estimate is recognized as a lower bound.

Added test case for "bar" data set for the runtime optimizer.  This query and data set ran into the lower bound estimate problem.  The change in this commit fixed the query.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3974 8f7bc0f5-282e-0410-95e3-8d296e9bb460