Merge CHANGE_SET_BRANCH to trunk [r3608:HEAD].
Note: There is a problem in TestChangeSets when run with TestBigdataSailWithQuads.  The test needs to be modified in order to not run the TM test variants when in quads mode. https://sourceforge.net/apps/trac/bigdata/ticket/166 has been amended to note this issue which will be fixed in the trunk.



git-svn-id: https://svn.code.sf.net/p/bigdata/code/trunk@3979 8f7bc0f5-282e-0410-95e3-8d296e9bb460