Worked through the split of a bucket and the split of a bucket where the address space must be doubled first.
The next step is to handle cases where splitting the bucket once does not result in a bucket with sufficient space for the new key. There are actually two cases here: (1) the hash codes of the keys are distinct, so if we double the address space enough times the insert will succeed; (2) the hash codes of the keys are identical, so no amount of expansion of the address space will permit the insert to succeed and an overflow page must be used.  For (1) we can also chose to use an overflow page in order to prevent run away expansion of the address space.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3988 8f7bc0f5-282e-0410-95e3-8d296e9bb460