Working through a problem with SubqueryOp leading to non-termination of some queries in BSBM.  This appears to be fixed at this checkpoint.  I am not sure yet if the fix was the -XX:+UseMembar JVM argument (I was running with JDK 1.6.0_17 which can loose monitor wake up signals) or the modifications to SubqueryOp.
I have made several changes to improve the trapping of "normal" query termination exceptions (InterruptedException, BufferClosedException, and ClosedByInterruptException).  All of these can arise in response to an interrupt triggered when a LIMIT is satisfied on the query.

The RWStore and WORMStore properties files were modified to turn on the new query evaluation strategy impl which handles optional join groups.  The data need to be loaded with that option enabled and also with dataTime inlining enabled.

Added more reporting capabilities to the NanoSparqlServer, including reporting of the BOPs for the currently executing queries.  This required adding a public method to IQueryClient to report on the UUIDs of the active IRunningQuery instances and making getRunningQuery(UUID) public so we can inspect the query by its UUID.

BlockingBuffer was generating a log message even when it would not log the message.  This is fixed.

RWStore was logging an error in getData() when there are several non-error reasons why that method could throw an exception (an interrupt leading to a ClosedByInterrupt exception is the main one).

Modified SOp2BOpUtility to layer the slice over the subquery op and not the other way around.

The current BSBM performance status is excellent for the reduced query mix (without Q5 or Q6).  Q5 appears to have a bad join plan which causes high CPU utilization.  Q6 is visiting much more data than is otherwise required in order to satisfy a regex without a prefix anchor.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4101 8f7bc0f5-282e-0410-95e3-8d296e9bb460