Added method to return an int[] of the ids for predicates to BOpUtility.
Added method to return the set of constraints attached to the last join in a join path to PartitionedJoinGroup and (a few) unit tests for that method.

Modified BOpBase#toString() to write out everything if an operand is an IValueExpression for better printing of bops.

Modified the QUADS mode SAIL proxy test suites to invoke the TestNestedOptionals and TestNestedUnions test suites.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4152 8f7bc0f5-282e-0410-95e3-8d296e9bb460