Some optimizations of heap churn (the bop deep copy semantics are now shallow copy semantics, which is sufficient to maintain their immutable contract).
Created a canonicalizing factory and applied it to the BigdataValueFactoryImpl, which is already shared for a given namespace.  I am still looking at how to apply this to the LexiconRelation to increase sharing of the term cache.

Added private loggers for the NanoHTTP class hierarchy.

Added test w/o shared variables to TestPipelineJoin.

Interned various strings which are used as annotation names.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4262 8f7bc0f5-282e-0410-95e3-8d296e9bb460