I have renamed the RDF aware Constraint class to SPARQLConstraint to avoid confusion with the version which is NOT aware of SPARQL evaluation semantics (esp, type errors).
I have added some optimizations to MathBOp and RangeBOp designed to provide a fast path if the left argument evaluations to null and to defer heap allocations until we know that the  RangeBOp can be fully evaluated.

I have modified SPOPredicate#asBound(...) to trap type errors arising from attempts to evaluate a RangeBOp when some variable(s) are not bound.  I am not sure that this is the right thing to do, but it allows the RTO to run.  It may be that the underlying problem is making the PartitionedJoinGroup aware of the RangeBOp.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4275 8f7bc0f5-282e-0410-95e3-8d296e9bb460