Bug fix for https://sourceforge.net/apps/trac/bigdata/ticket/252 (inline dateTime causes error with full tx).
I modified the LexiconRelation to always use an unisolated index for mutable views.  The lexicon relation uses an eventually consistent strategy to write on the TERM2ID and ID2TERM indices and does not rely on transactional isolation.  AbstractRelation#getIndex/3 is used to automatically wrap the unisolated views of those indices with an UnisolatedReadWriteIndex class, which uses a RecurrentReadWriteLock to control access to the unisolated indices.

I have added com.bigdata.rdf.sail.TextTxCreate.  It verifies that a Sail can be created which supports transactional isolation both with and without inline date times.  I have removed the sample code for CreateSailUsingInlineDateTimes in favor of this unit test.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4277 8f7bc0f5-282e-0410-95e3-8d296e9bb460