Merged LARGE_LITERALS_REFACTOR to QUADS_QUERY_BRANCH [r4174:r4387].
SVN reported a 23 file conflicts and 115 tree conflicts. None of the file conflicts was a real conflict.  In all cases the version from the LARGE_LITERAL_REFACTOR was accepted.  

This change set includes:
-	HashCollisionUtility, which is used to test various designs for large literal handling.
-	New performance counters in several classes (Journal, BTree, etc).
-	The MemStore, which is based on the MemoryManager.
-	Various fixes to the MemoryManager, but there are still issues remaining.  See the notes at the bottom of the MemoryManager class for details.
-	Support for “raw records” in the B+Tree.  This is a configurable option which permits you to specify the maximum byte length of the value associated with a tuple in the B+Tree before it is written as raw record on the backing file rather than inlined into the leaf.

The following tests were run before committing the merge:
-	TestBigdataSailWithQuads – 11 expected test failures.
-	TestMemoryManager - ok
-	TestLocalTripleStore - ok
-	TestRWJournal
o	resolved test_restartSafe_writeOne_noCommit where the exception thrown had been changed and was also being wrapped as a RuntimeException, but the test had not been updated.  This test failure was already present in the QUADS branch.
o	Removed explicit case to ResourceManager in AbstractTask which was causing several RWStore tests to fail. This bug was introduced within the LARGE_LITERALS_REFACTOR branch when I added getIndexCounters() to the IResourceManager API.
o	Bug fix to com.bigdata.journal.TestRootBlockView.test_ctor_correctRejection.  This has been failing for some time.  The problem is that the commit record addr and commit record index addr MUST both be NULL (0L) if the commit counter is zero.  This was a unit test problem.  I’ve added several new cases to the unit test and fixed the one that was failing.  Also changed all references to System.err into conditional logging @ INFO.
-	TestWORMJournal – Ok.
-	TestMemStore – Ok.
-	TestAll (BTree) – Ok (except for 4 known failures for the index segment which have been around since the merge from the JOURNAL_HA_BRANCH).


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4388 8f7bc0f5-282e-0410-95e3-8d296e9bb460