Continued work on the htree. At this point I believe that I have all of the relevant bit math working with appropriate unit tests (except for maskOffLSB).
This commit includes a refactoring of the internal interfaces for the B+Tree to support some of the differences between the BTree and the HTree.

The next steps on the HTree are to write the recursive descent and tuple management logic, followed by handling splits in the buddy buckets and buddy hash tables and the gradual deepening of the hash tree.

Some of the mechanisms for persistence have been incorporated into the hash tree, but there needs to be a reconcile between the HTree and the BTree at some point so they can share more code (abstract classes) and interfaces for generalized persistent indices.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4463 8f7bc0f5-282e-0410-95e3-8d296e9bb460