Added HTree support for lookupFirst() and lookupAll(), insert() with split of the bucket (but not yet introduction of a new directory level), and contains() along with basic unit tests for the same.
Next steps will be:

1. A split of a bucket page consisting of a single buddy bucket (globalDepth==localDepth).  This requires the introduction of a new directory page.

2. Remove.

3. More detailed tests, stress tests, etc.

Note that the data structure does not yet support persistence.  I am working to get the core logic for the HTree maintence correct and tested first.  I can then bring in the additional infrastructure to support incremental writes, checkpoint records, etc.  Much of this is already stubbed out, but not yet implemented.  

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4477 8f7bc0f5-282e-0410-95e3-8d296e9bb460