Fixed some asserts in the HTree constructor.
Introduced a simpler example based on inserting 1,2,3,4,.... into the HTree.  This example will make it possible to explore introducing a new directory page into the hash tree without having to deal with all the edge cases at once.  The old example is still present in the test suite (and the worksheet) and I will revisit it after the basics of directory splitting are in place.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4512 8f7bc0f5-282e-0410-95e3-8d296e9bb460