Working to resolve [1].
I've added versions of this test with truth maintenance enabled (for triples and sids modes).

I've modified TestSids?, which runs before this test in the TestBigdataSailWithSids? test suite, to improve its finally clauses handling the open connection and the open sail.

I've also checked TestTxCreate?, which runs before this test in the TestBigdataSailWithSids? and the TestBigdataSailWithQuads? test suites.

There were ambiguous invocations of BigdataSailConnection#isOpen?() in two derived classes. I have modified to those to test the 'open' field instead.

I've added versions of this test with truth maintenance enabled (for triples and sids modes).

I've modified TestSids?, which runs before this test in the TestBigdataSailWithSids? test suite, to improve its finally clauses handling the open connection and the open sail.

I've also checked TestTxCreate?, which runs before this test in the TestBigdataSailWithSids? and the TestBigdataSailWithQuads? test suites.

There were ambiguous invocations of BigdataSailConnection#isOpen?() in two derived classes. I have modified to those to test the 'open' field instead.

Renamed 'open' fields on BigdataSail? and BigdataSailConnection? as openSail and openConn to help identify incorrect inheritance problems.

Renamed assertOpen() methods as assertOpenConn() and assertOpenSail().

Renamed assertWritable() as assertWritableConn()

Modified toString() on the connection classes to show whether the connection is open.

Made the anonymous BigdataSailConnection? classes into named inner classes in an attempt to prevent future ambiguity in the inheritance patterns.

Note: It would be far better if BigdataSailConnection? were a static class since that would remove the opportunity entirely for incorrect inheritance patterns.

TestSearchQuery - the Journal was not closing the TemporaryStoreFactory for a normal close() but only in destroyResources().  I moved that code into _close(), which is now overridden by the Journal.

TestSearchQuery#test_restart() - improved nested try/finally to close connections.

[1] https://sourceforge.net/apps/trac/bigdata/ticket/315 (TestRollbacksTx does not release buffers/close stores)

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4600 8f7bc0f5-282e-0410-95e3-8d296e9bb460