This turned out to be a relatively simple issue.  The ant 'bundle' task now flattens the directory structure of the jars.  This was causing a problem for the cluster deployment since the jini jars were divided into three directories (jini/lib, jini/lib-dl, and jini-lib-ext). The solution was simply to divide the "bundle" task into two "copy" operations.  One flattens most of the jars, but does not handle the jars which are specific to the cluster deployment.  The other copies the jars for the cluster deployment, maintaining their directory hierarchy.
This issue only effected the 'ant-install-artifact' task and service startup when using "bigdata 'start'".  It would not have effected the alternative install procedure which is based on the REL.<version>.tgz artifact since that does not rely on the 'bundle' target.

I also modified bigdataStandalone.config to not override the class path for zookeeper.  Doing so is only asking for problems since then you have to maintain the right version information for the zookeeper dependencies in the config file as well.

See https://sourceforge.net/apps/trac/bigdata/ticket/314

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4606 8f7bc0f5-282e-0410-95e3-8d296e9bb460