Bug fix for [1].
This appears to be a bug in Constant#equals(Object o). The code was written to allow comparison with IConstantOrVariable. However, IVariable#get() always throws an UnsupportedOperationException? since it is not possible to obtain the asBound value of a variable without reference to a binding set.

I have modified Constant#equals() to return false unless the passed Object is another IConstant. I also added a fast code path for tests against self. This is sufficient to have the unit test pass. I am now running through the test suites for the SAIL to verify that this change has not broken anything else.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4613 8f7bc0f5-282e-0410-95e3-8d296e9bb460