Another check point on the TERMS refactor.
Making good progress on replacing the old lexicon indices with the new TERMS index.  I've worked through the vocabulary, axioms, lexicon, and spo test suites for the triple store.

I worked through the bnode write on the TERMS index when NOT using told bnodes.  In this mode, bnodes never unify with the terms already in the index and we always add a new bnode having the same hash code whose collision counter is the range count of the collision bucket for that VTE and hash code.

Most of the problems with the test suite come down to the old TermId(VTE,termId) constructor.  I've introduced a factory pattern for this.  I'm working through callers of that constructor and misc methods which reach out of the old TERM2ID or ID2TERM index.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@4642 8f7bc0f5-282e-0410-95e3-8d296e9bb460