Fixed lexicon joins (but I still need to write the test suite for them).
Added substantially to the unit tests for the IVs and uncovered several problems with their Comparable implementations (some of which still need to be fixed).

Epoch extension now supports negatives (time existed before the epoch).

Removed the TERMID_BITS_TO_REVERSE stuff.

Cleaned up much of the LexiconRelation and friends.


git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@4689 8f7bc0f5-282e-0410-95e3-8d296e9bb460