Added the original SPARQL and parsed SPARQL operator tree in their own sections of the page.
Modified to stream the explanation and status page results rather than buffering in memory.

Removed the XMLBuilder and HTMLBuilder constructor variants which did not require the caller to provide the target on which to write the data.  The XMLBuilder#toString() assumed that the Writer was a StringWriter. That assumption was not compatible with the use of the class when building a servlet response.

See https://sourceforge.net/apps/trac/bigdata/ticket/331

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4726 8f7bc0f5-282e-0410-95e3-8d296e9bb460