There is now only a single parser invocation per NanoSparqlServer request. In order to achieve this I had to bundle a modified version of the openrdf SPARQLParser, which is called "BigdataSPARQLParser".  The modified version extracts the query hints and the query type (ASK, CONSTRUCT, DESCRIBE, or SELECT) and reports them out via an interface which is implemented by the returned ParsedQuery (IBigdataParsedQuery).  The logic hacked into the BigdataSailRespositoryConnection (for query hints) and into QueryType (for SELECT, ASK, DESCRIBE or CONSTRUCT) has been removed.
The NanoSparqlServer was also modified to report a BAD_REQUEST if there is an error in the SPARQL and to include the text of the syntax error message in the response line.

I'm not sure whether we should register the BigdataSparqlParser over the SPARQLParser.  The bigdata variant is used by the NanoSparqlServer and the BigdataSailResponsitoryConnection.  It is pretty much a requirement for setting up a bigdata sail query, but there might not be any reason to explicitly override the existing SPARQL parser registration.

See https://sourceforge.net/apps/trac/bigdata/ticket/336

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4750 8f7bc0f5-282e-0410-95e3-8d296e9bb460