Actually, the hot spot is probably the synchronization internally on the writeRetentionQueue. {{{         if (!writeRetentionQueue.add(node)) { }}}
For a read-only B+Tree, this is already configured by newWriteRetentionQueue() to use a HardReferenceQueueWithBatchingUpdates, which uses a lock amortization strategy (striped locks).  We could look at using either true thread-local buffers as a higher performance lock amortization technique or at changing the parameters used to configure the HardReferenceQueueWithBatchingUpdates instance.

I've refactored AbstractBTree slightly since we now know in the constructor whether or not the B+Tree view is read-only.  There is now a protected readOnly field on AbstractBTree and this field is used directly from within AbstractBTree#touch() to decide whether or not additional synchronization is required.  I have also refactored touch() to pull out a doSyncTouch() method as a complement to the existing doTouch().  The doSyncTouch() method is for a mutable B+Tree.  The doTouch() method is directly invoked for a read-only B+Tree.  I am going to use this distinction in the code paths to further isolate the hot spot.

See https://sourceforge.net/apps/trac/bigdata/ticket/201

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4764 8f7bc0f5-282e-0410-95e3-8d296e9bb460