Updated the zookeeper bundled dependency to 3.3.3 from 3.2.1.  The specific version is now parameterized through build.properties. Build.xml was modified to support that version parameterization.
Updated the tuprolog dependency from 2.1.1 to 2.2 and bundled the 2p jar rather than the tuprolog jar.  The 2p jar includes the CLI and GUI interfaces and is more useful for interactive work with prolog.  The tuprolog jar contains the core of the prolog implementation is might be all that it required for some applications, but it is not really enough for application development.

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4769 8f7bc0f5-282e-0410-95e3-8d296e9bb460