The 3rd party operation integration feature is done.  See TestSetBindingSets for an example of how to make use of this feature.
I've made the Property for the QueryHints null if there are no query hints and modified the various classes which look things up in that Properties object.

I've backed out the changes which expose this facility in the BigdataSailQuery interface.  For now, the mechanism is available to people who directly invoke BigdataSailConnection#evaluate(...).  I prefer this approach with its minimum propagation of the API change until we get some feedback on the utility of this integration.

See https://sourceforge.net/apps/trac/bigdata/ticket/267

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@4790 8f7bc0f5-282e-0410-95e3-8d296e9bb460