I have modified BigdataValueFactoryImpl per the change recommended above.
I have integrated the unit test into the test suite for TestBigdataSailWithQuads.

I have extended the test suite for the BigdataValueFactoryImpl to verify the behavior when a DummyIV is used.

These changes have been applied to the 1.0.0 release branch and to the current development branch.

See https://sourceforge.net/apps/trac/bigdata/ticket/348

git-svn-id: https://svn.code.sf.net/p/bigdata/code/branches/BIGDATA_RELEASE_1_0_0@4890 8f7bc0f5-282e-0410-95e3-8d296e9bb460